Added aws upload
